Finished Model, just need to comment

+++ 4005Project/src/Event.java
@@ -2,7 +2,7 @@
 public interface Event {
     
 	 public EventTypes getEventType();
-	 public int getEventfTime();
-	 public int getEventsTime();
+	 public double getEventfTime();
+	 public double getEventsTime();
 	 public String toString();
 }
